create audioProps file and transfer all code
