+++ src/Sequencer/Sequencer.js
@@ -3,120 +3,18 @@ import * as Tone from "tone";
 import * as d3 from "d3-random";
 import Sketch from "react-p5";
 import Square from "../Square/Square";
-
-const initialPattern = [
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  //   [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  //   [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  //   [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  //   [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-];
-
-const bassSynth = new Tone.MembraneSynth().toDestination();
-bassSynth.volume.value = -6;
-
-const dist = new Tone.Distortion(0.2).toDestination();
-const snareSynth = new Tone.NoiseSynth({
-  volume: -18,
-  noise: {
-    type: "white",
-    playbackRate: 3,
-  },
-  envelope: {
-    attack: 0.001,
-    decay: 0.2,
-    sustain: 0,
-    release: 0.05,
-  },
-}).connect(dist);
-
-const hiHatSynth = new Tone.MetalSynth({
-  envelope: {
-    attack: 0.001,
-    decay: 0.05,
-    sustain: 0,
-    release: 0.03,
-  },
-  resonance: 8000,
-}).connect(dist);
-hiHatSynth.volume.value = -6;
-
-const pluckSynth = new Tone.PluckSynth().toDestination();
-pluckSynth.volume.value = -6;
-
-const notes = ["A3", "C4", "D4", "E4", "G4", "A4"];
-const initialPolySynthPattern = [
-  [1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0],
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-];
-
-const reverb = new Tone.Reverb(3).toDestination();
-reverb.wet.value = 0.5;
-const delay = new Tone.PingPongDelay("3.5n", 0.3).connect(reverb);
-delay.wet.value = 0.3;
-const polySynth = new Tone.PolySynth(Tone.DuoSynth).connect(delay);
-polySynth.set({
-  harmonicity: 3,
-  detune: -1200,
-  voice0: {
-    envelope: {
-      attack: 0.001,
-      decay: 0.5,
-      sustain: 0,
-      release: 0.01,
-    },
-  },
-  voice1: {
-    envelope: {
-      attack: 0.001,
-      decay: 0.5,
-      sustain: 0,
-      release: 0.01,
-    },
-  },
-});
-polySynth.volume.value = -24;
-
-const initialMonoSynthPattern = [
-  [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  //   [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-  //   [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-];
-
-const filter = new Tone.Filter().toDestination();
-const monoSynth = new Tone.MonoSynth({
-  envelope: {
-    attack: 0.01,
-    decay: 0.4,
-    sustain: 0,
-    release: 0.01,
-  },
-}).connect(filter);
-monoSynth.volume.value = -15;
-const lfo = new Tone.LFO("2m", 100, 4000).start().connect(filter.frequency);
+import audioProps from "../audioProps";
 
 const Sequencer = () => {
   const [clicked, setClicked] = useState(false);
   const [playState, setPlayState] = useState(Tone.Transport.state);
   const [activeColumn, setColumn] = useState(0);
-  const [pattern, updatePattern] = useState(initialPattern);
+  const [pattern, updatePattern] = useState(audioProps.initialPattern);
   const [polySynthPattern, setPolySynthPattern] = useState(
-    initialPolySynthPattern
+    audioProps.initialPolySynthPattern
   );
   const [monoSynthPattern, setMonoSynthPattern] = useState(
-    initialMonoSynthPattern
+    audioProps.initialMonoSynthPattern
   );
 
   useEffect(
@@ -130,13 +28,25 @@ const Sequencer = () => {
             // If active
             if (row[col] && noteIndex === 0) {
               // Play based on which row
-              bassSynth.triggerAttackRelease("C0", "16nn", time + "+0.1");
+              audioProps.bassSynth.triggerAttackRelease(
+                "C0",
+                "16nn",
+                time + "+0.1"
+              );
             } else if (row[col] && noteIndex === 1) {
-              snareSynth.triggerAttackRelease("8n", time + "+0.1");
+              audioProps.snareSynth.triggerAttackRelease("8n", time + "+0.1");
             } else if (row[col] && noteIndex === 2) {
-              hiHatSynth.triggerAttackRelease("C1", "16n", time + "+0.1");
+              audioProps.hiHatSynth.triggerAttackRelease(
+                "C1",
+                "16n",
+                time + "+0.1"
+              );
             } else if (row[col] && noteIndex === 3) {
-              pluckSynth.triggerAttackRelease("C4", "16n", time + "+0.1");
+              audioProps.pluckSynth.triggerAttackRelease(
+                "C4",
+                "16n",
+                time + "+0.1"
+              );
             }
           });
         },
@@ -161,7 +71,7 @@ const Sequencer = () => {
     const chance = d3.randomUniform(0.5, 1.5)();
 
     // Loop through and create some random on/off values
-    const row = psPatternCopy[tuneData % notes.length];
+    const row = psPatternCopy[tuneData % audioProps.notes.length];
     tuneData++;
     for (let x = 0; x < row.length; x++) {
       row[x] = Math.abs(d3.randomNormal()()) > chance ? 1 : 0;
@@ -183,10 +93,14 @@ const Sequencer = () => {
         let notesToPlay = [];
         polySynthPattern.forEach((row, noteIndex) => {
           if (row[col]) {
-            notesToPlay.push(notes[noteIndex]);
+            notesToPlay.push(audioProps.notes[noteIndex]);
           }
         });
-        polySynth.triggerAttackRelease(notesToPlay, "16n", time + "+0.1");
+        audioProps.polySynth.triggerAttackRelease(
+          notesToPlay,
+          "16n",
+          time + "+0.1"
+        );
       },
       [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15],
       "16n"
@@ -200,13 +114,13 @@ const Sequencer = () => {
       (time, col) => {
         if (steps128 < 64) {
           steps128++;
-          monoSynth.triggerAttackRelease("C2", "16n", time + "+0.1");
+          audioProps.monoSynth.triggerAttackRelease("C2", "16n", time + "+0.1");
         } else if (steps128 >= 64 && steps128 < 127) {
           steps128++;
-          monoSynth.triggerAttackRelease("A1", "16n", time + "+0.1");
+          audioProps.monoSynth.triggerAttackRelease("A1", "16n", time + "+0.1");
         } else {
           steps128 = 0;
-          monoSynth.triggerAttackRelease("A1", "16n", time + "+0.1");
+          audioProps.monoSynth.triggerAttackRelease("A1", "16n", time + "+0.1");
         }
       },
       [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15],
